auth-service/pyproject.toml
[project]
name = "auth-serivce"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.13"
readme = "README.md"
dependencies = [
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings>=2.9.1,<3.0.0",
    "alembic (>=1.16.1,<2.0.0)",
    "ruff>=0.11.11",
    "fastapi>=0.115.12",
    "greenlet>=3.2.2",
    "pytest>=8.3.5",
    "uvicorn>=0.34.2",
    "pydantic[email]>=2.11.5",
    "passlib[bcrypt]>=1.7.4",
    "types-passlib>=1.7.7.20250516",
    "pyjwt>=2.10.1",
]

[tool.uv]
package = false

[tool.ruff]
exclude = ["alembic"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
plugins = ["pydantic.mypy"]
